--- frontend/pages/hosts/HostDetailsPage/SoftwareTab/EmptySoftware/_styles.scss
@@ -1,54 +1,0 @@
-.empty-software {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: $pad-xxxlarge;
-  margin-bottom: $pad-xxxlarge;
-
-  &__inner {
-    display: flex;
-    flex-direction: row;
-
-    h1 {
-      font-size: $small;
-      font-weight: $bold;
-      margin-bottom: $pad-medium;
-    }
-
-    img {
-      width: 176px;
-      margin-right: 34px;
-    }
-
-    p {
-      color: $core-fleet-black;
-      font-weight: $regular;
-      font-size: $x-small;
-      margin: 0;
-    }
-
-    p.learn-more {
-      margin-top: $pad-medium;
-    }
-
-    a {
-      color: $core-vibrant-blue;
-      font-size: $x-small;
-      font-weight: $bold;
-      text-decoration: none;
-      margin-left: $pad-xsmall;
-    }
-
-    img {
-      width: 12px;
-      height: 12px;
-      margin-left: 6px;
-    }
-  }
-
-  &__empty-filter-results {
-    display: flex;
-    flex-direction: column;
-    width: 350px;
-  }
-}
