--- frontend/pages/hosts/HostDetailsPage/SoftwareTab/EmptySoftware/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./EmptySoftware";
