--- frontend/pages/hosts/HostDetailsPage/SoftwareTab/SoftwareVulnCount/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SoftwareVulnCount";
