+++ frontend/pages/hosts/SoftwareTab/EmptySoftware/_styles.scss
