+++ frontend/pages/hosts/SoftwareTab/SoftwareTab.tsx
