+++ frontend/pages/hosts/SoftwareTab/SoftwareVulnCount/SoftwareVulnCount.tsx
