+++ frontend/pages/hosts/SoftwareTab/SoftwareVulnCount/_styles.scss
