+++ frontend/pages/hosts/SoftwareTab/SoftwareVulnCount/index.ts
